Renames and testing pipenv
