+++ docs/PillowCommands.txt
